requirements.txt
Flask>=2.3,<4.0
Werkzeug>=2.3
torch>=2.0
torchvision>=0.15
Pillow>=10.0
numpy>=1.24
